added a player controls
