+++ lib/player.rb
@@ -1,16 +1,9 @@
-require_relative '../lib/game_logic.rb'
-require_relative '../bin/main.rb'
 # instruction to play game
-# class PlayGame
-#   def game_play
-#     game = GameLogic.new
-#     game.user until game.game_over?
-#     if game.win?
-#       @winner = game.winner
-#       @winner = @name
-#       puts "Congratulations #{@name}!"
-#     elsif game.draw?
-#       puts 'Good try, You draw!'
-#     end
-#   end
-# end
+class Player
+  attr_writer :name
+  def player_check(name)
+    @name = name
+    player2 if @name == player1
+  end
+  self
+end
